Move test_coordinator tests to test_init Add test for duplicate config entry
